requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2.3
django-cors-middleware==1.5.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-gis==0.17
idna==2.10
jwcrypto==0.8
oauthlib==3.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
